+++ packages/server/src/server/agent/providers/claude/agent.test.ts
@@ -610,6 +610,14 @@ describe("ClaudeAgentSession features", () => {
       client.listFeatures({
         provider: "claude",
         cwd: process.cwd(),
+        model: "claude-opus-4-8-20260101",
+      }),
+    ).resolves.toEqual([expect.objectContaining({ id: "fast_mode", value: false })]);
+
+    await expect(
+      client.listFeatures({
+        provider: "claude",
+        cwd: process.cwd(),
         model: "claude-sonnet-4-6",
       }),
     ).resolves.toEqual([]);
